% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingModel.R
\docType{methods}
\name{BayesianMissingAggregate}
\alias{BayesianMissingAggregate}
\title{BayesianMissingAggregate}
\usage{
BayesianMissingAggregate(obj, ...)
}
\arguments{
\item{obj}{A QFeatures instance.}

\item{...}{Additional arguments passed to control MCMC settings.}

\item{i}{A string or integer specifying the element of 'obj' that contains the log abundance intensities that will be modeled.}

\item{fcol}{A string  that denotes feature variable of assay ‘i’ defining how to summarise the features.}

\item{gcol}{A string or integer specifying which column of colData(obj) denotes sample -> group assignments}

\item{contrastPrefix}{A string that prefixes model results names.}

\item{comparisons}{A contrast matrix, generated by `limma::makeContrasts`,
specifying which pairwise group comparisons to perform.}

\item{threshold}{Numeric. The parameter that determines
which missingness mechanism model to use. If `threshold = 1`,
the logistic model is used; If `threshold = 0`, the truncated normal model is used.}

\item{parallel}{Logical. If `TRUE`, the analysis for each protein is
parallelized across multiple CPU cores.}

\item{ROPE}{A numeric vector of length 2 specifying the lower and upper bounds of the
Region of Practical Equivalence (ROPE). Values within this interval are considered
practically equivalent to the null value (e.g., `c(-0.1, 0.1)` for log fold change).}

\item{n.adapt}{Integer. Number of iterations for the adaptive phase.}

\item{burn.in}{An integer denoting number of burn-in steps.}

\item{n.iter}{Integer. Number of iterations after burn-in.}

\item{n.chains}{Integer. Number of MCMC chains to run.}

\item{mcmcDiag}{Logical. If `TRUE`, the output includes additional MCMC diagnostics.}
}
\value{
Original QFeatures object 'obj' with model results attached.
}
\description{
QFeatures integration, runs model fitting for all specified comparisons and adds them to existing QFeatures object.
}
\seealso{
\code{\link{BayesMissingModel}}
}
