% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{f_beta}
\alias{f_beta}
\title{Helper function to retrieve beta based on sample mean
A helper function that maps group mean value to its
corresponding rate parameter for the inverse-gamma prior on
within-group variance.}
\usage{
f_beta(mu_val, beta, a, b)
}
\arguments{
\item{mu_val}{A numeric group mean value.}

\item{beta}{A vector of beta parameters returned by `sigma_jp2params`.}

\item{a}{The lower bound of the intensity bins returned by `sigma_jp2params`.}

\item{b}{The upper bound of the intensity bins returned by `sigma_jp2params`.}
}
\value{
The beta parameter for the inverse-gamma prior corresponding to the
  intensity bin that `mu_val` falls into.
}
\description{
Helper function to retrieve beta based on sample mean
A helper function that maps group mean value to its
corresponding rate parameter for the inverse-gamma prior on
within-group variance.
}
