% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{generate_overdispersed_inits}
\alias{generate_overdispersed_inits}
\title{Generate overdispersed initial values for MCMC chains
Creates a list of initial values for JAGS model.}
\usage{
generate_overdispersed_inits(
  sample_mean_1,
  sample_mean_2,
  sample_var_1,
  sample_var_2,
  n_chains,
  spread_factor = 3
)
}
\arguments{
\item{sample_mean_1}{Sample mean of the first group.}

\item{sample_mean_2}{Sample mean of the second group.}

\item{sample_var_1}{Sample variance of the first group.}

\item{sample_var_2}{Sample variance of the second group.}

\item{n_chains}{The number of MCMC chains.}

\item{spread_factor}{A multiplier to control the degree of overdispersion from the sample means.}
}
\value{
A list of length `n_chains`.
}
\description{
Generate overdispersed initial values for MCMC chains
Creates a list of initial values for JAGS model.
}
\keyword{internal}
