% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{HDIofMCMC}
\alias{HDIofMCMC}
\title{Function for computing limits of HDI's: Computes highest density interval from a sample of representative values, estimated as shortest credible interval.
This function was written by John K. Kruschke for his book
"Doing Bayesian Data Analysis" (2nd edition).}
\usage{
HDIofMCMC(sampleVec, credMass = 0.95)
}
\arguments{
\item{sampleVec}{a vector of representative values from a probability distribution.}

\item{credMass}{is a scalar between 0 and 1, indicating the mass within the credible interval that is to be estimated.}
}
\value{
a vector containing the limits of the HDI
}
\description{
Function for computing limits of HDI's: Computes highest density interval from a sample of representative values, estimated as shortest credible interval.
This function was written by John K. Kruschke for his book
"Doing Bayesian Data Analysis" (2nd edition).
}
\references{
Kruschke, J. K. (2015). Doing Bayesian Data Analysis:
  A Tutorial with R, JAGS, and Stan (2nd ed.). Academic Press.
}
\keyword{internal}
