% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{setupContrasts}
\alias{setupContrasts}
\title{Subset data and make specific contrast
Prepares a dataset for differential analysis between two specified groups.
Subsets the global data matrix to the corresponding samples and optionally
filter to include only proteins that contain missing values
.}
\usage{
setupContrasts(overall_distri, comparisons, contrast, group, filter4NAs)
}
\arguments{
\item{overall_distri}{A numeric matrix or data.frame of log2 intensities. Rows correspond to proteins
(with protein IDs as rownames), and columns correspond to samples. Each row must
contain at least one non-NA value.}

\item{comparisons}{A matrix showing which groups to compare for each contrast, output by limma::makeContrasts.}

\item{contrast}{A string specifying which two groups are compared.}

\item{group}{A factor vector specifying the group assignment for each sample.}

\item{filter4NAs}{Logical. If `TRUE`, filters the dataset to retain only rows (proteins)
that contain at least one missing value.}
}
\value{
A list containing the data for the contrast:
  \item{subset_data}{The subset to the two groups being compared.}
  \item{group_subset}{The group factor subset to the two groups.}
  \item{groups_to_compare}{The names of the two groups being compared.}
  \item{cp}{A vector of minimum values per protein from the full dataset,
  used as a cutoff.}
}
\description{
Subset data and make specific contrast
Prepares a dataset for differential analysis between two specified groups.
Subsets the global data matrix to the corresponding samples and optionally
filter to include only proteins that contain missing values
.
}
\keyword{internal}
