% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{zeroState}
\alias{zeroState}
\title{Calculate data-level hyperparameters
Estimates the global mode of the data (`mu_0`) and the coefficients (`gamma0`, `gamma1`)
for a logistic regression that models the probability of a value being observed
as a function of its intensity.}
\usage{
zeroState(overall_distri)
}
\arguments{
\item{overall_distri}{A numeric matrix or data.frame of log2 intensities. Rows correspond to proteins
(with protein IDs as rownames), and columns correspond to samples. Each row must
contain at least one non-NA value.}
}
\value{
A named numeric vector of hyperparameters:
  \item{mu_0}{The estimated global mode of the data distribution.}
  \item{gamma0}{The intercept of logistic model.}
  \item{gamma1}{The slope of logistic model.}
}
\description{
Calculate data-level hyperparameters
Estimates the global mode of the data (`mu_0`) and the coefficients (`gamma0`, `gamma1`)
for a logistic regression that models the probability of a value being observed
as a function of its intensity.
}
