% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingModel.R
\name{BayesMissingModel}
\alias{BayesMissingModel}
\title{Bayesian Differential Expression Analysis for Proteomics
This is the primary function for performing Bayesian differential expression analysis
on proteomics datasets.}
\usage{
BayesMissingModel(
  values,
  groups,
  comparisons,
  filter4NAs = FALSE,
  threshold = 0,
  parallel = TRUE,
  ROPE = c(-0.2, 0.2),
  n.adapt = 1000,
  burn.in = 500,
  n.iter = 10000,
  n.chains = 2,
  mcmcDiag = FALSE
)
}
\arguments{
\item{values}{A numeric matrix or data.frame of log2 intensities. Rows correspond to proteins
(with protein IDs as rownames), and columns correspond to samples. Each row must
contain at least one non-NA value.}

\item{groups}{A factor vector specifying the group assignment for each sample.}

\item{comparisons}{A contrast matrix, generated by `limma::makeContrasts`,
specifying which pairwise group comparisons to perform.}

\item{filter4NAs}{Logical. If `TRUE`, filters the dataset to retain only rows (proteins)
that contain at least one missing value.}

\item{threshold}{Numeric. The parameter that determines
which missingness mechanism model to use. If `threshold = 1`,
the logistic model is used; If `threshold = 0`, the truncated normal model is used.}

\item{parallel}{Logical. If `TRUE`, the analysis for each protein is
parallelized across multiple CPU cores.}

\item{ROPE}{A numeric vector of length 2 specifying the lower and upper bounds of the
Region of Practical Equivalence (ROPE). Values within this interval are considered
practically equivalent to the null value (e.g., `c(-0.1, 0.1)` for log fold change).}

\item{n.adapt}{Integer. Number of iterations for the adaptive phase.}

\item{burn.in}{An integer denoting number of burn-in steps.}

\item{n.iter}{Integer. Number of iterations after burn-in.}

\item{n.chains}{Integer. Number of MCMC chains to run.}

\item{mcmcDiag}{Logical. If `TRUE`, the output includes additional MCMC diagnostics.}
}
\value{
A list of tables containing model results for each contrast specified in comparisons.
}
\description{
Bayesian Differential Expression Analysis for Proteomics
This is the primary function for performing Bayesian differential expression analysis
on proteomics datasets.
}
\examples{

my_data <- matrix(rnorm(1200), nrow = 300, ncol = 4)
rownames(my_data) <- paste0("Protein_", 1:300)
colnames(my_data) <- paste0("Sample_", 1:4)
groups_factor <- factor(c("Control", "Control", "Treatment", "Treatment"))


# Create a contrast matrix
my_contrasts <- limma::makeContrasts(
Treatment_vs_Control = Treatment - Control,
levels = groups_factor
)

# Run the analysis
output <- BayesMissingModel(
values = my_data,
groups = groups_factor,
comparisons = my_contrasts,
parallel = FALSE
)


}
\seealso{
\code{\link{plotPost}} for visualizing the posterior distribution of a single protein.
}
