% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{plotPost}
\alias{plotPost}
\title{Plot Posterior Distribution}
\usage{
plotPost(
  values,
  id,
  groups,
  contrast,
  threshold = 0,
  cenTend = c("mode", "median", "mean")[1],
  compVal = 0,
  ROPE = c(-0.2, 0.2),
  n.adapt = 500,
  burn.in = 500,
  n.iter = 5000,
  n.chains = 2,
  credMass = 0.95,
  HDItextPlace = 0.7,
  xlab = NULL,
  xlim = NULL,
  yaxt = NULL,
  ylab = NULL,
  main = NULL,
  cex = NULL,
  cex.lab = NULL,
  col = NULL,
  border = NULL,
  showCurve = FALSE,
  breaks = NULL,
  ...
)
}
\arguments{
\item{values}{A numeric matrix or data.frame of log2 intensities. Rows correspond to proteins
(with protein IDs as rownames), and columns correspond to samples. Each row must
contain at least one non-NA value.}

\item{id}{A character string specifying the protein ID to plot.}
}
\value{
A posterior distribution plot.
}
\description{
Creates a plot of posterior distribution with HDI and other statistics.
}
\details{
This function is modified from the \code{plotPost} function written by John K. Kruschke for his book
"Doing Bayesian Data Analysis" (2nd edition).
Modifications include: Users can directly input the data, protein ID, and other information to get the posterior distribution.
}
\examples{
\dontrun{
# Assuming 'my_data' is a matrix and 'groups' is a factor
plotPost(
  values = my_data,
  id = "Protein_ID",
  groups = groups,
  contrast = "Treatment - Control",
  compVal = 0,
  ROPE = c(-0.1, 0.1)
)
}
}
\references{
Kruschke, J. K. (2015). Doing Bayesian Data Analysis:
A Tutorial with R, JAGS, and Stan (2nd ed.). Academic Press.
}
\author{
John K. Kruschke (original `plotPost` function), Mengchun Li (port and integration with `missBayes`)
}
