% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_row.R
\name{runModel}
\alias{runModel}
\title{Execute the Bayesian model for differential expression analysis
This function conducts the full Bayesian analysis for a specified contrast.}
\usage{
runModel(
  contrast_data,
  sigma_p2p,
  sigma_jp2p,
  zS,
  sigma_1,
  default_data,
  n.chains,
  burn.in,
  n.iter,
  n.adapt,
  parallel,
  mcmcDiag,
  threshold,
  ROPE
)
}
\value{
A data frame with one row per protein and columns containing the
  summarized posterior statistics. The row names correspond to the protein IDs
  from the input data.
}
\description{
Execute the Bayesian model for differential expression analysis
This function conducts the full Bayesian analysis for a specified contrast.
}
\note{
Discription of arguments see BayesMissingModel.
}
\seealso{
\code{\link{BayesMissingModel}}
}
