% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{sigma_p2params}
\alias{sigma_p2params}
\title{Estimate hyperparameters for the between-group variance prior}
\usage{
sigma_p2params(overall_distri, group)
}
\arguments{
\item{overall_distri}{A numeric matrix or data.frame of log2 intensities. Rows correspond to proteins
(with protein IDs as rownames), and columns correspond to samples. Each row must
contain at least one non-NA value.}

\item{group}{A factor vector specifying the group assignment for each sample.}
}
\value{
A list containing:
  \item{alpha_p}{Shape of the inverse-gamma prior.}
  \item{beta_p}{Rate of the inverse-gamma prior.}
  \item{r}{Number of replicates per group.}
  \item{group_means}{Matrix of mean value for each protein in each group.}
}
\description{
This function models the between-group variance.
}
\keyword{internal}
